Solved several bugs about the spatial dynamics of cells.
The position and angle were not updated correctly when running simulation in command line mode. The reason was that the GUI called the Cell::updateGraphicsCell(const double time) function that broadcastes the position and angles from CellBase class to all the other GraphicsCell classes. Now we also call the update function in the main simulator.